CRT-52 rename: auth.controller 패키지 이름을 auth.api로 변경
